+++ trobo/src/main/java/com/allstate/trobo/dao/impl/JdbcTripsheetRepository.java
@@ -59,8 +59,8 @@ public class JdbcTripsheetRepository implements TripSheetRepository {
 			Address address = new Address();
 			address.setId(rs.getLong("id"));
 			address.setAddressLine(rs.getString("address_line"));
-			address.setLongitude(rs.getDouble("longitude"));
-			address.setLatitude(rs.getDouble("latitude"));
+			address.setLongitude(rs.getBigDecimal("longitude"));
+			address.setLatitude(rs.getBigDecimal("latitude"));
 			pickUpPoint.setAddress(address);
 			pickUpPoint.setNumberOfEmployees(rs.getInt("empCount"));
 			return pickUpPoint;
